+++ sdk-core/src/main/java/io/nem/sdk/model/transaction/MultisigAccountModificationTransactionFactory.java
@@ -27,8 +27,8 @@ import org.apache.commons.lang3.Validate;
 public class MultisigAccountModificationTransactionFactory extends
     TransactionFactory<MultisigAccountModificationTransaction> {
 
-    private final int minApprovalDelta;
-    private final int minRemovalDelta;
+    private final byte minApprovalDelta;
+    private final byte minRemovalDelta;
     private final List<MultisigCosignatoryModification> modifications;
 
     private MultisigAccountModificationTransactionFactory(
@@ -67,7 +67,7 @@ public class MultisigAccountModificationTransactionFactory extends
      *
      * @return byte
      */
-    public int getMinApprovalDelta() {
+    public byte getMinApprovalDelta() {
         return minApprovalDelta;
     }
 
@@ -77,7 +77,7 @@ public class MultisigAccountModificationTransactionFactory extends
      *
      * @return byte
      */
-    public int getMinRemovalDelta() {
+    public byte getMinRemovalDelta() {
         return minRemovalDelta;
     }
